Exception Handler

+++ src/main/java/com/devsuperior/dsecommerce/services/ProductService.java
@@ -1,15 +1,20 @@
 package com.devsuperior.dsecommerce.services;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.devsuperior.dsecommerce.dto.ProductDTO;
 import com.devsuperior.dsecommerce.entities.Product;
 import com.devsuperior.dsecommerce.repositories.ProductRepository;
+import com.devsuperior.dsecommerce.services.expections.DatabaseExpection;
 import com.devsuperior.dsecommerce.services.expections.ResourceNotFoundExpection;
+
+import jakarta.persistence.EntityNotFoundException;
 
 
 @Service
@@ -43,14 +48,32 @@ public class ProductService {
 	
 	@Transactional
 	public ProductDTO upload(Long id, ProductDTO dto) {
-		Product entity = repository.getReferenceById(id);
-		copyDtoToEntity(dto, entity);
-		entity = repository.save(entity);
-		return new ProductDTO(entity);
+		
+		try {
+			Product entity = repository.getReferenceById(id);
+			copyDtoToEntity(dto, entity);
+			entity = repository.save(entity);
+			return new ProductDTO(entity);
+		}
+		catch(EntityNotFoundException e){
+			throw new ResourceNotFoundExpection("Recurso não encontrato");
+		}
+		
 	}
 	
+	@Transactional(propagation = Propagation.SUPPORTS)
 	public void delete(Long id) {
-		repository.deleteById(id);
+		
+		if (!repository.existsById(id)) {
+			throw new ResourceNotFoundExpection("Recurso não encontrado");
+		}
+		try {
+	        	repository.deleteById(id);    		
+		}
+	    catch (DataIntegrityViolationException e) {
+	        	throw new DatabaseExpection("Falha de integridade referencial");
+	   	}
+		
 	}
 	
 	private void copyDtoToEntity(ProductDTO dto, Product entity) {
